requirements.txt
discord==2.3.2
python-dotenv==1.0.1
yt-dlp==2024.4.9